third_party/WebKit/LayoutTests/editing/pasteboard/cut-paste-formatting-tag.html

<!doctype HTML>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>

test(() => assert_selection(
  '<div contenteditable style="width: 300px; height: 250px; border: 1px solid black"><div><b>^<br></b></div><b>foo|</b></div>',
  selection => {
    selection.document.execCommand('cut');
    selection.document.execCommand('paste');
  },
  '<div contenteditable style="width: 300px; height: 250px; border: 1px solid black"><br><div><b>foo|</b></div></div>'),
  'Cut all tags and paste them');

</script>
